Moved Mocks to Sextant.Mocks
