+++ src/Sextant.Mocks/Mocks/PageViewModelMock.cs
@@ -3,7 +3,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
-namespace Sextant.Testing
+namespace Sextant.Mocks
 {
     /// <summary>
     /// A mock of a page view model.
