+++ src/Sextant.Mocks/Sextant.Mocks.csproj
@@ -2,10 +2,10 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
+    <NoWarn>$(NoWarn);1591;CA1707;SA1633;CA1034</NoWarn>
+    <IsPackable>false</IsPackable>
   </PropertyGroup>
-  <ItemGroup>
-    <None Include="$(MSBuildThisFileDirectory)ApiGeneratorGlobalSuppressions.cs" />
-  </ItemGroup>
+
   <ItemGroup>
     <ProjectReference Include="..\Sextant\Sextant.csproj" />
   </ItemGroup>
