--- src/Sextant.Testing/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace Sextant.Testing
-{
-    public class Class1
-    {
-    }
-}
